Fixes Model utils (#34)
* Fixes Model utils

* fixes smoke test

* fixes accuracy

* tries adding tests

* fixes setup and docs
