+++ docs/templates/index.md
@@ -1,21 +1,3 @@
-# Vision
+# Quickvision
 
-- Computer Vision for Autonomous Vehicles.
-
-<div align="center">
-
-[![GitHub issues](https://img.shields.io/github/issues/VAUV/vision)](https://github.com/VAUV/vision/issues)
-[![GitHub forks](https://img.shields.io/github/forks/VAUV/vision)](https://github.com/VAUV/vision/network)
-[![GitHub stars](https://img.shields.io/github/stars/VAUV/vision)](https://github.com/VAUV/vision/stargazers)
-[![GitHub license](https://img.shields.io/github/license/VAUV/vision)](https://github.com/VAUV/vision)
-[![codecov](https://codecov.io/gh/VAUV/vision/branch/master/graph/badge.svg?token=KM2TLB42R8)](undefined)
-
-![PEP8](https://github.com/VAUV/vision/workflows/Check%20Code%20formatting/badge.svg)
-![CI Tests](https://github.com/VAUV/vision/workflows/CI%20Tests/badge.svg)
-![Docs](https://github.com/VAUV/vision/workflows/Deploy%20mkdocs/badge.svg)
-![PyPi Release](https://github.com/VAUV/vision/workflows/PyPi%20Release/badge.svg)
-
-</div>
-
-
-End to End training of various Computer Vision Models
+- Faster Computer Vision
