+++ quickvision/losses/detr_loss.py
@@ -8,7 +8,8 @@ from torch import nn
 from torchvision.ops.misc import interpolate
 from torchvision.ops import generalized_box_iou, box_convert
 from torchvision.ops import sigmoid_focal_loss
-from quickvision.models.model_utils import accuracy, nested_tensor_from_tensor_list
+from quickvision.utils import nested_tensor_from_tensor_list
+from quickvision.metrics import accuracy
 from quickvision.losses.segmentation import dice_loss
 import torch.nn.functional as F
