+++ quickvision/metrics/__init__.py
@@ -1,0 +1,1 @@
+from quickvision.metrics.classification import accuracy
