+++ quickvision/utils/torch_utils.py
@@ -1,0 +1,35 @@
+import os
+import torch
+import numpy as np
+import random
+import torchvision
+from typing import Optional, List
+from torch import Tensor
+
+__all__ = ["set_debug_apis", "seed_everything", "print_size_of_model"]
+
+
+def set_debug_apis(state: bool = False):
+    torch.autograd.profiler.profile(enabled=state)
+    torch.autograd.profiler.emit_nvtx(enabled=state)
+    torch.autograd.set_detect_anomaly(mode=state)
+
+
+def seed_everything(seed):
+    """
+    Makes code deterministic using a given seed.
+    Internally sets all seeds of torch, numpy and random.
+    """
+    random.seed(seed)
+    os.environ["PYTHONHASHSEED"] = str(seed)
+    np.random.seed(seed)
+    torch.manual_seed(seed)
+    torch.cuda.manual_seed(seed)
+    torch.backends.cudnn.deterministic = True
+    torch.backends.cudnn.benchmark = True
+
+
+def print_size_of_model(model):
+    torch.save(model.state_dict(), "temp.p")
+    print("Size (MB):", os.path.getsize("temp.p") / 1e6)
+    os.remove("temp.p")
