+++ tests/test_detr.py
@@ -7,7 +7,6 @@ import pytorch_lightning as pl
 from torch_utils import im2tensor
 from quickvision.models.detection import detr
 from quickvision.models.detection.detr import create_detr_backbone
-from quickvision.models import model_utils
 from quickvision.losses import detr_loss
 from quickvision.models.detection.detr import engine
 from dataset_utils import DummyDetectionDataset
